# src/main/java/com/eaglebank/dto/response/ListTransactionsResponse.java
package com.eaglebank.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ListTransactionsResponse {
    private List<TransactionResponse> transactions;
}